@@ -1,4 +1,4 @@
-﻿﻿using System;
+﻿using System;
 using System.Collections.Generic;
 using System.Diagnostics;
 using System.Diagnostics.CodeAnalysis;
@@ -63,7 +63,7 @@ namespace PKHeX.Core.AutoMod
                 template.Version = dest.Game;
 
             template.ApplySetDetails(set);
-            template.SetRecordFlags(Array.Empty<ushort>()); // Validate TR/MS moves for the encounter
+            template.SetRecordFlags([]); // Validate TR/MS moves for the encounter
 
             if (template.Species == (ushort)Species.Unown) // Force unown form on template
                 template.Form = set.Form;
@@ -309,7 +309,7 @@ namespace PKHeX.Core.AutoMod
 
                 GameVersion[] result;
                 if (value.IsValidSavedVersion())
-                    result = new GameVersion[] { value };
+                    result = [value];
                 else
                     result = GameUtil.GameVersions.Where(z => value.Contains(z)).ToArray();
 
@@ -367,9 +367,9 @@ namespace PKHeX.Core.AutoMod
                 mutate = (LanguageID)idx;
             if (AllowTrainerOverride && regen.HasTrainerSettings && regen.Trainer != null)
                 return regen.Trainer.MutateLanguage(mutate, ver);
-            if (UseTrainerData)
-                return TrainerSettings.GetSavedTrainerData(ver, gen).MutateLanguage(mutate, ver);
-            return TrainerSettings.DefaultFallback(ver, regen.Extra.Language);
+            return UseTrainerData
+                ? TrainerSettings.GetSavedTrainerData(ver, gen).MutateLanguage(mutate, ver)
+                : TrainerSettings.DefaultFallback(ver, regen.Extra.Language);
         }
 
         /// <summary>
@@ -498,10 +498,7 @@ namespace PKHeX.Core.AutoMod
                 requested = rt.Regen.Extra.Alpha;
 
             // Requested alpha but encounter isn't an alpha
-            if (enc is not IAlphaReadOnly a)
-                return !requested;
-
-            return a.IsAlpha == requested;
+            return enc is not IAlphaReadOnly a ? !requested : a.IsAlpha == requested;
         }
 
         public static bool IsRequestedShinyValid(IBattleTemplate set, IEncounterable enc)
@@ -544,9 +541,7 @@ namespace PKHeX.Core.AutoMod
                 return true;
             if (enc is IStaticCorrelation8b s && s.GetRequirement(pk) == StaticCorrelation8bRequirement.MustHave)
                 return true;
-            if (enc is EncounterEgg && GameVersion.BDSP.Contains(enc.Version))
-                return true;
-            return false;
+            return enc is EncounterEgg && GameVersion.BDSP.Contains(enc.Version);
         }
 
         /// <summary>
@@ -649,7 +644,7 @@ namespace PKHeX.Core.AutoMod
                 // check for mixed->fixed gender incompatibility by checking the gender of the original species
                 if (SpeciesCategory.IsFixedGenderFromDual(pk.Species) && pk.Gender != 2) // shedinja
                 {
-                    pk.Gender = EntityGender.GetFromPID(new LegalInfo(pk, new List<CheckResult>()).EncounterMatch.Species,pk.EncryptionConstant);
+                    pk.Gender = EntityGender.GetFromPID(new LegalInfo(pk, []).EncounterMatch.Species, pk.EncryptionConstant);
                     // genderValid = true; already true if we reach here
                 }
             }
@@ -688,9 +683,7 @@ namespace PKHeX.Core.AutoMod
         /// <returns></returns>
         public static Func<int, int, int> CompetitiveMarking(PKM pk)
         {
-            if (pk.Format < 7)
-                return GetSimpleMarking;
-            return GetComplexMarking;
+            return pk.Format < 7 ? GetSimpleMarking : GetComplexMarking;
 
             static int GetSimpleMarking(int val, int _) => val == 31 ? 1 : 0;
             static int GetComplexMarking(int val, int _) =>
@@ -1241,7 +1234,7 @@ namespace PKHeX.Core.AutoMod
                         continue;
                 }
 
-                if (ratio != PersonalInfo.RatioMagicMale && ratio != PersonalInfo.RatioMagicFemale  && ratio != PersonalInfo.RatioMagicGenderless)
+                if (ratio != PersonalInfo.RatioMagicMale && ratio != PersonalInfo.RatioMagicFemale && ratio != PersonalInfo.RatioMagicGenderless)
                 {
                     var gender_roll = rng.NextUInt(252) + 1;
                     var fin_gender = gender_roll < ratio ? 1 : 0;
@@ -1312,11 +1305,9 @@ namespace PKHeX.Core.AutoMod
                 return false;
             if ((uint)template.Gender < 2 && template.Gender != pk.Gender) // match gender
                 return false;
-            if (template.Form != pk.Form && !FormInfo.IsFormChangeable( pk.Species, pk.Form, template.Form, EntityContext.Gen9, pk.Context)) // match form -- Toxtricity etc
+            if (template.Form != pk.Form && !FormInfo.IsFormChangeable(pk.Species, pk.Form, template.Form, EntityContext.Gen9, pk.Context)) // match form -- Toxtricity etc
                 return false;
-            if (template.Shiny != pk.IsShiny)
-                return false;
-            return true;
+            return template.Shiny == pk.IsShiny;
         }
 
         /// <summary>
@@ -1457,36 +1448,36 @@ namespace PKHeX.Core.AutoMod
                 pk.Egg_Location = Locations.LinkTrade4; // todo: really shouldn't be doing this, don't modify pkm
                 return PIDType.Method_1;
             }
-            var info = new LegalInfo(pk, new List<CheckResult>());
+            var info = new LegalInfo(pk, []);
             EncounterFinder.FindVerifiedEncounter(pk, info);
             return pk.Generation switch
             {
                 3 => info.EncounterMatch switch
+                {
+                    WC3 g => g.Method,
+                    EncounterStatic3 when pk.Version == (int)GameVersion.CXD => PIDType.CXD,
+                    EncounterStatic3Colo when pk.Version == (int)GameVersion.CXD => PIDType.CXD,
+                    EncounterStatic3XD when pk.Version == (int)GameVersion.CXD => PIDType.CXD,
+                    EncounterStatic3 => pk.Version switch
                     {
-                        WC3 g => g.Method,
-                        EncounterStatic3 when pk.Version == (int)GameVersion.CXD => PIDType.CXD,
-                        EncounterStatic3Colo when pk.Version == (int)GameVersion.CXD => PIDType.CXD,
-                        EncounterStatic3XD when pk.Version == (int)GameVersion.CXD => PIDType.CXD,
-                        EncounterStatic3 => pk.Version switch
-                            {
-                                (int)GameVersion.E => PIDType.Method_1,
-                                (int)GameVersion.FR or (int)GameVersion.LG => PIDType.Method_1, // roamer glitch
-                                _ => PIDType.Method_1,
-                            },
-
-                        EncounterSlot3 when pk.Version == (int)GameVersion.CXD => PIDType.PokeSpot,
-                        EncounterSlot3XD when pk.Version == (int)GameVersion.CXD => PIDType.PokeSpot,
-                        EncounterSlot3 => pk.Species == (int)Species.Unown ? PIDType.Method_1_Unown : PIDType.Method_1,
-                        _ => PIDType.None,
+                        (int)GameVersion.E => PIDType.Method_1,
+                        (int)GameVersion.FR or (int)GameVersion.LG => PIDType.Method_1, // roamer glitch
+                        _ => PIDType.Method_1,
                     },
+
+                    EncounterSlot3 when pk.Version == (int)GameVersion.CXD => PIDType.PokeSpot,
+                    EncounterSlot3XD when pk.Version == (int)GameVersion.CXD => PIDType.PokeSpot,
+                    EncounterSlot3 => pk.Species == (int)Species.Unown ? PIDType.Method_1_Unown : PIDType.Method_1,
+                    _ => PIDType.None,
+                },
 
                 4 => info.EncounterMatch switch
-                    {
-                        EncounterStatic4Pokewalker => PIDType.Pokewalker,
-                        EncounterStatic4 s => (s.Shiny == Shiny.Always ? PIDType.ChainShiny : PIDType.Method_1),
-                        PGT => PIDType.Method_1,
-                        _ => PIDType.None,
-                    },
+                {
+                    EncounterStatic4Pokewalker => PIDType.Pokewalker,
+                    EncounterStatic4 s => (s.Shiny == Shiny.Always ? PIDType.ChainShiny : PIDType.Method_1),
+                    PGT => PIDType.Method_1,
+                    _ => PIDType.None,
+                },
 
                 _ => PIDType.None,
             };
@@ -1727,18 +1718,12 @@ namespace PKHeX.Core.AutoMod
         /// <summary>
         /// Async Related actions for global timer.
         /// </summary>
-        public class AsyncLegalizationResult
+        public class AsyncLegalizationResult(
+            PKM pk,
+            LegalizationResult res)
         {
-            public readonly PKM Created;
-            public readonly LegalizationResult Status;
-
-            public AsyncLegalizationResult(
-                PKM pk,
-                LegalizationResult res)
-            {
-                Created = pk;
-                Status = res;
-            }
+            public readonly PKM Created = pk;
+            public readonly LegalizationResult Status = res;
         }
 
         private static async Task<AsyncLegalizationResult?>? TimeoutAfter(
@@ -1765,9 +1750,7 @@ namespace PKHeX.Core.AutoMod
             };
 
             var res = group.GetVersionsWithinRange(versionlist.ToArray());
-            if (res.Length > 0)
-                return res;
-            return new[] { version };
+            return res.Length > 0 ? res : ([version]);
         }
     }
 }
