@@ -22,15 +22,14 @@ namespace PKHeX.Core.AutoMod
     {
         public static string GetMessage(this AutoModErrorCode code)
         {
-            if (code.IsSilent())
-                return string.Empty;
-
-            return code switch
-            {
-                AutoModErrorCode.NotEnoughSpace => "Not enough space in the box.",
-                AutoModErrorCode.InvalidLines => "Invalid lines detected.",
-                _ => string.Empty,
-            };
+            return code.IsSilent()
+                ? string.Empty
+                : code switch
+                {
+                    AutoModErrorCode.NotEnoughSpace => "Not enough space in the box.",
+                    AutoModErrorCode.InvalidLines => "Invalid lines detected.",
+                    _ => string.Empty,
+                };
         }
 
         public static bool IsSilent(this AutoModErrorCode code) =>
