@@ -56,7 +56,7 @@ namespace PKHeX.Core.AutoMod
                 sb.AppendLine(RegenUtil.GetSummary(Batch));
             if (EncounterFilters.Any())
                 sb.AppendLine(RegenUtil.GetSummary(EncounterFilters));
-            if (VersionFilters.Count() > 0)
+            if (VersionFilters.Count > 0)
                 sb.AppendLine(RegenUtil.GetSummary(VersionFilters));
             return sb.ToString();
         }
