@@ -10,10 +10,7 @@ namespace PKHeX.Core.AutoMod
             if (gen is 9)
             {
                 // Scatterbug and Spewpa must be Fancy
-                if (
-                    set.Species == (ushort)Species.Scatterbug
-                    || set.Species == (ushort)Species.Spewpa
-                )
+                if (set.Species == (ushort)Species.Scatterbug || set.Species == (ushort)Species.Spewpa)
                     set.Form = 18;
                 return;
             }
@@ -33,18 +30,18 @@ namespace PKHeX.Core.AutoMod
             {
                 case (ushort)Species.Zacian:
                 case (ushort)Species.Zamazenta:
-                {
-                    // Behemoth Blade and Behemoth Bash -> Iron Head
-                    if (!set.Moves.Contains((ushort)781) && !set.Moves.Contains((ushort)782))
-                        return;
-
-                    for (int i = 0; i < set.Moves.Length; i++)
                     {
-                        if (set.Moves[i] is 781 or 782)
-                            set.Moves[i] = 442;
+                        // Behemoth Blade and Behemoth Bash -> Iron Head
+                        if (!set.Moves.Contains((ushort)781) && !set.Moves.Contains((ushort)782))
+                            return;
+
+                        for (int i = 0; i < set.Moves.Length; i++)
+                        {
+                            if (set.Moves[i] is 781 or 782)
+                                set.Moves[i] = 442;
+                        }
+                        break;
                     }
-                    break;
-                }
             }
         }
 
@@ -54,10 +51,7 @@ namespace PKHeX.Core.AutoMod
         /// <param name="set"></param>
         public static void SanitizeTeraTypes(this RegenTemplate set)
         {
-            if (
-                set.Species == (int)Species.Ogerpon
-                && !TeraTypeUtil.IsValidOgerpon((byte)set.TeraType, set.Form)
-            )
+            if (set.Species == (int)Species.Ogerpon && !TeraTypeUtil.IsValidOgerpon((byte)set.TeraType, set.Form))
                 set.TeraType = ShowdownEdits.GetValidOpergonTeraType(set.Form);
         }
 
