@@ -32,6 +32,8 @@ namespace PKHeX.Core.Enhancements
                 "Complete"
             };
 
+        private static readonly string[] separator = new[] { "\n" };
+
         public static bool IsInvalidForm(string form) =>
             form.Contains("Mega") || InvalidFormes.Contains(form);
 
@@ -45,7 +47,7 @@ namespace PKHeX.Core.Enhancements
             paste = paste.Trim(); // Remove White Spaces
             if (IsTeamBackup(paste))
                 return ShowdownTeamSet.GetTeams(paste).SelectMany(z => z.Team).ToList();
-            var lines = paste.Split(new[] { "\n" }, StringSplitOptions.None);
+            var lines = paste.Split(separator, StringSplitOptions.None);
             return ShowdownParsing.GetShowdownSets(lines).ToList();
         }
 
@@ -59,7 +61,7 @@ namespace PKHeX.Core.Enhancements
             source = source.Trim();
             if (IsTeamBackup(source))
                 return true;
-            string[] stringSeparators = { "\n\r" };
+            string[] stringSeparators = ["\n\r"];
 
             var result = source.Split(stringSeparators, StringSplitOptions.None);
             return new ShowdownSet(result[0]).Species > 0;
