@@ -68,9 +68,7 @@ namespace PKHeX.Core.Enhancements
         {
             if (url.Contains("pokepast.es/"))
                 return PasteSource.PokePaste;
-            if (url.Contains("pastebin.com/"))
-                return PasteSource.Pastebin;
-            return PasteSource.None;
+            return url.Contains("pastebin.com/") ? PasteSource.Pastebin : PasteSource.None;
         }
 
         private string GetRawURL(string url)
@@ -92,17 +90,17 @@ namespace PKHeX.Core.Enhancements
             switch (Source)
             {
                 case PasteSource.PokePaste:
-                {
-                    var url = URL.Replace("/raw", "");
-                    GetFromPokePaste(url);
-                    return;
-                }
+                    {
+                        var url = URL.Replace("/raw", "");
+                        GetFromPokePaste(url);
+                        return;
+                    }
                 case PasteSource.Pastebin:
-                {
-                    var url = URL.Replace("/raw/", "/");
-                    GetFromPasteBin(url);
-                    return;
-                }
+                    {
+                        var url = URL.Replace("/raw/", "/");
+                        GetFromPasteBin(url);
+                        return;
+                    }
                 default:
                     return; // This should never happen
             }
