@@ -83,15 +83,15 @@ namespace PKHeX.Core.Injection
 
         private byte[] ReadData(ulong offset, RWMethod method)
         {
-            if (Bot.com is not ICommunicatorNX nx)
-                return Bot.ReadOffset(offset);
-            return method switch
-            {
-                RWMethod.Heap => Bot.ReadOffset(offset),
-                RWMethod.Main => nx.ReadBytesMain(offset, Bot.SlotSize),
-                RWMethod.Absolute => nx.ReadBytesAbsolute(offset, Bot.SlotSize),
-                _ => Bot.ReadOffset(offset),
-            };
+            return Bot.com is not ICommunicatorNX nx
+                ? Bot.ReadOffset(offset)
+                : method switch
+                {
+                    RWMethod.Heap => Bot.ReadOffset(offset),
+                    RWMethod.Main => nx.ReadBytesMain(offset, Bot.SlotSize),
+                    RWMethod.Absolute => nx.ReadBytesAbsolute(offset, Bot.SlotSize),
+                    _ => Bot.ReadOffset(offset),
+                };
         }
 
         public byte[] ReadRAM(ulong offset, int size) => Bot.com.ReadBytes(offset, size);
